Signal creation specifies min and max by pointer.
No longer uses INFINITY as a constant to indicate no min/max.  Instead
takes a pointer to a value.  Also provide setters for min and max, and
a general msig_new function, called by msig_float and msig_int.  Also
provide missing msig_free(), and it's called by mdev_free().

Provide a couple of shorter macros to make casting to
mapper_signal_value_t (and a pointer to it) a bit easier.