In signal handler, provide the value as a void* instead of mapper_signal_value_t*.
In fact, the max external made the exact error that this change is
intended to prevent!  (i.e., indexing into an array of
mapper_signal_value_t instead of an array of float/int/etc.)  So,
shows that it's probably better this way.

We need more testing code for vector signals.